fix wrong type data

+++ common.go
@@ -81,7 +81,7 @@ type Position struct {
 	ShortOrderSize               float64 `json:"shortOrderSize"`
 	Side                         string  `json:"side"`
 	Size                         float64 `json:"size"`
-	UnrealizedPnl                int     `json:"unrealizedPnl"`
+	UnrealizedPnl                float64 `json:"unrealizedPnl"`
 	CollateralUsed               float64 `json:"collateralUsed"`
 }
